[ZAK] Fix test_valid_code_compiles: main() should return i64 not i32
- compile_and_run_zeta expects i64 return value
- Test had main() -> i32 which caused type mismatch error
- Changed to main() -> i64 to match expected return type
- All tests should now pass

+++ tests/error_handling_integration.rs
@@ -96,7 +96,7 @@ fn test_error_recovery_multiple_errors() {
 #[test]
 fn test_valid_code_compiles() {
     let code = r#"
-        fn main() -> i32 {
+        fn main() -> i64 {
             let x = 42;
             let y = 10;
             x + y
